Implement a partially complete UpsertTimerActivity
 The activity creates a new TimerGroup upon creation, and sets the internal timerGroupId, and fetches all the timers with that id

+++ app/src/main/java/io/takano/timeous/timers/TimerDao.java
@@ -22,10 +22,10 @@ public interface TimerDao {
     void delete(Timer timer);
 
     @Query("DELETE FROM timers WHERE group_id=:groupId")
-    void deleteTimersInGroup(Integer groupId);
+    void deleteTimersInGroup(Long groupId);
 
     @Query("SELECT * FROM timers WHERE group_id=:groupId")
-    LiveData<List<Timer>> getTimersInGroup(Integer groupId);
+    LiveData<List<Timer>> getTimersInGroup(Long groupId);
 
     @Query("SELECT * FROM timers")
     LiveData<List<Timer>> getAllTimers();
